+++ src/version2/jQL.ts
@@ -96,8 +96,8 @@ export class JQL {
       url: '/rest/api/2/jql/autocompletedata',
       method: 'POST',
       data: {
-        projectIds: parameters?.projectIds,
         includeCollapsedFields: parameters?.includeCollapsedFields,
+        projectIds: parameters?.projectIds,
       },
     };
 
@@ -304,6 +304,53 @@ export class JQL {
       method: 'POST',
       data: {
         queries: parameters?.queries,
+      },
+    };
+
+    return this.client.sendRequest(config, callback);
+  }
+
+  async getPrecomputations<T = Models.PageJqlFunctionPrecomputation>(
+    parameters: Parameters.GetPrecomputations | undefined,
+    callback: Callback<T>
+  ): Promise<void>;
+  async getPrecomputations<T = Models.PageJqlFunctionPrecomputation>(
+    parameters?: Parameters.GetPrecomputations,
+    callback?: never
+  ): Promise<T>;
+  async getPrecomputations<T = Models.PageJqlFunctionPrecomputation>(
+    parameters?: Parameters.GetPrecomputations,
+    callback?: Callback<T>,
+  ): Promise<void | T> {
+    const config: RequestConfig = {
+      url: '/rest/api/2/jql/function/computation',
+      method: 'GET',
+      params: {
+        functionKey: parameters?.functionKey,
+        startAt: parameters?.startAt,
+        maxResults: parameters?.maxResults,
+        orderBy: parameters?.orderBy,
+        filter: parameters?.filter,
+      },
+    };
+
+    return this.client.sendRequest(config, callback);
+  }
+
+  async updatePrecomputations<T = unknown>(
+    parameters: Parameters.UpdatePrecomputations,
+    callback: Callback<T>
+  ): Promise<void>;
+  async updatePrecomputations<T = unknown>(parameters: Parameters.UpdatePrecomputations, callback?: never): Promise<T>;
+  async updatePrecomputations<T = unknown>(
+    parameters: Parameters.UpdatePrecomputations,
+    callback?: Callback<T>,
+  ): Promise<void | T> {
+    const config: RequestConfig = {
+      url: '/rest/api/2/jql/function/computation',
+      method: 'POST',
+      data: {
+        values: parameters.values,
       },
     };