+++ src/version2/models/jqlFunctionPrecomputationUpdateRequest.ts
@@ -1,0 +1,6 @@
+import { JqlFunctionPrecomputationUpdate } from './jqlFunctionPrecomputationUpdate';
+
+/** List of pairs (id and value) for precomputation updates. */
+export interface JqlFunctionPrecomputationUpdateRequest {
+  values: JqlFunctionPrecomputationUpdate[];
+}