+++ src/version2/models/updateNotificationSchemeDetails.ts
@@ -1,0 +1,7 @@
+/** Details of a notification scheme. */
+export interface UpdateNotificationSchemeDetails {
+  /** The description of the notification scheme. */
+  description?: string;
+  /** The name of the notification scheme. Must be unique. */
+  name?: string;
+}