+++ src/version2/parameters/findGroups.ts
@@ -1,7 +1,7 @@
 export interface FindGroups {
   /**
-   * This parameter is deprecated, setting it does not affect the results. To find groups containing a particular user,
-   * use [Get user groups](#api-rest-api-2-user-groups-get).
+   * @deprecated This parameter is deprecated, setting it does not affect the results. To find groups containing a
+   *   particular user, use [Get user groups](#api-rest-api-2-user-groups-get).
    */
   accountId?: string;
   /** The string to find in group names. */
@@ -11,7 +11,7 @@ export interface FindGroups {
    * the result. To exclude multiple groups, provide an ampersand-separated list. For example,
    * `exclude=group1&exclude=group2`. This parameter cannot be used with the `excludeGroupIds` parameter.
    */
-  exclude?: string[];
+  exclude?: string | string[];
   /**
    * A group ID to exclude from the result. To exclude multiple groups, provide an ampersand-separated list. For
    * example, `excludeId=group1-id&excludeId=group2-id`. This parameter cannot be used with the `excludeGroups`
@@ -23,10 +23,12 @@ export interface FindGroups {
    * property `jira.ajax.autocomplete.limit`.
    */
   maxResults?: number;
+  /** Whether the search for groups should be case-insensitive. */
+  caseInsensitive?: boolean;
   /**
-   * This parameter is no longer available. See the [deprecation
-   * notice](https://developer.atlassian.com/cloud/jira/platform/deprecation-notice-user-privacy-api-migration-guide/)
-   * for details.
+   * @deprecated This parameter is no longer available. See the [deprecation
+   *   notice](https://developer.atlassian.com/cloud/jira/platform/deprecation-notice-user-privacy-api-migration-guide/)
+   *   for details.
    */
   userName?: string;
 }