+++ src/version2/parameters/getPrecomputations.ts
@@ -1,0 +1,7 @@
+export interface GetPrecomputations {
+  functionKey?: string[];
+  startAt?: number;
+  maxResults?: number;
+  orderBy?: string;
+  filter?: string;
+}