+++ src/version2/projectComponents.ts
@@ -16,7 +16,7 @@ export class ProjectComponents {
    * _Administer projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project in which the
    * component is created or _Administer Jira_ [global permission](https://confluence.atlassian.com/x/x4dKLg).
    */
-  async createComponent<T = Models.Component>(
+  async createComponent<T = Models.ProjectComponent>(
     parameters: Parameters.CreateComponent,
     callback: Callback<T>
   ): Promise<void>;
@@ -29,8 +29,11 @@ export class ProjectComponents {
    * _Administer projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project in which the
    * component is created or _Administer Jira_ [global permission](https://confluence.atlassian.com/x/x4dKLg).
    */
-  async createComponent<T = Models.Component>(parameters: Parameters.CreateComponent, callback?: never): Promise<T>;
-  async createComponent<T = Models.Component>(
+  async createComponent<T = Models.ProjectComponent>(
+    parameters: Parameters.CreateComponent,
+    callback?: never
+  ): Promise<T>;
+  async createComponent<T = Models.ProjectComponent>(
     parameters: Parameters.CreateComponent,
     callback?: Callback<T>,
   ): Promise<void | T> {
@@ -38,20 +41,20 @@ export class ProjectComponents {
       url: '/rest/api/2/component',
       method: 'POST',
       data: {
-        self: parameters.self,
-        id: parameters.id,
-        name: parameters.name,
-        description: parameters.description,
-        lead: parameters.lead,
-        leadUserName: parameters.leadUserName,
-        leadAccountId: parameters.leadAccountId,
-        assigneeType: parameters.assigneeType,
         assignee: parameters.assignee,
-        realAssigneeType: parameters.realAssigneeType,
-        realAssignee: parameters.realAssignee,
+        assigneeType: parameters.assigneeType,
+        description: parameters.description,
+        id: parameters.id,
         isAssigneeTypeValid: parameters.isAssigneeTypeValid,
+        lead: parameters.lead,
+        leadAccountId: parameters.leadAccountId,
+        leadUserName: parameters.leadUserName,
+        name: parameters.name,
         project: parameters.project,
         projectId: parameters.projectId,
+        realAssignee: parameters.realAssignee,
+        realAssigneeType: parameters.realAssigneeType,
+        self: parameters.self,
       },
     };
 
@@ -66,7 +69,7 @@ export class ProjectComponents {
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** _Browse
    * projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for project containing the component.
    */
-  async getComponent<T = Models.Component>(
+  async getComponent<T = Models.ProjectComponent>(
     parameters: Parameters.GetComponent | string,
     callback: Callback<T>
   ): Promise<void>;
@@ -78,8 +81,8 @@ export class ProjectComponents {
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** _Browse
    * projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for project containing the component.
    */
-  async getComponent<T = Models.Component>(parameters: Parameters.GetComponent | string, callback?: never): Promise<T>;
-  async getComponent<T = Models.Component>(
+  async getComponent<T = Models.ProjectComponent>(parameters: Parameters.GetComponent | string, callback?: never): Promise<T>;
+  async getComponent<T = Models.ProjectComponent>(
     parameters: Parameters.GetComponent | string,
     callback?: Callback<T>,
   ): Promise<void | T> {
@@ -103,7 +106,7 @@ export class ProjectComponents {
    * _Administer projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project containing
    * the component or _Administer Jira_ [global permission](https://confluence.atlassian.com/x/x4dKLg).
    */
-  async updateComponent<T = Models.Component>(
+  async updateComponent<T = Models.ProjectComponent>(
     parameters: Parameters.UpdateComponent,
     callback: Callback<T>
   ): Promise<void>;
@@ -117,8 +120,11 @@ export class ProjectComponents {
    * _Administer projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project containing
    * the component or _Administer Jira_ [global permission](https://confluence.atlassian.com/x/x4dKLg).
    */
-  async updateComponent<T = Models.Component>(parameters: Parameters.UpdateComponent, callback?: never): Promise<T>;
-  async updateComponent<T = Models.Component>(
+  async updateComponent<T = Models.ProjectComponent>(
+    parameters: Parameters.UpdateComponent,
+    callback?: never
+  ): Promise<T>;
+  async updateComponent<T = Models.ProjectComponent>(
     parameters: Parameters.UpdateComponent,
     callback?: Callback<T>,
   ): Promise<void | T> {
@@ -270,7 +276,7 @@ export class ProjectComponents {
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** _Browse
    * Projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project.
    */
-  async getProjectComponents<T = Models.Component[]>(
+  async getProjectComponents<T = Models.ProjectComponent[]>(
     parameters: Parameters.GetProjectComponents | string,
     callback: Callback<T>
   ): Promise<void>;
@@ -284,11 +290,11 @@ export class ProjectComponents {
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** _Browse
    * Projects_ [project permission](https://confluence.atlassian.com/x/yodKLg) for the project.
    */
-  async getProjectComponents<T = Models.Component[]>(
+  async getProjectComponents<T = Models.ProjectComponent[]>(
     parameters: Parameters.GetProjectComponents | string,
     callback?: never
   ): Promise<T>;
-  async getProjectComponents<T = Models.Component[]>(
+  async getProjectComponents<T = Models.ProjectComponent[]>(
     parameters: Parameters.GetProjectComponents | string,
     callback?: Callback<T>,
   ): Promise<void | T> {